refactor: share cloud reclaim lifecycle contracts
